Admin commands are in the new system

+++ src/main/kotlin/Bot.kt
@@ -9,8 +9,8 @@ import net.dv8tion.jda.api.requests.GatewayIntent
 
 class Bot(token: String) : ListenerAdapter() {
     var prefix = "."
-    val adminCommands = mutableMapOf<String, (msg: Message) -> Unit>()
-    val commands = mutableMapOf<String, (msg: Message) -> Unit>()
+    val commands = mutableListOf<Command>()
+    val commandsOld = mutableMapOf<String, (msg: Message) -> Unit>()
     val triggers = mutableMapOf<String, (msg: Message) -> Unit>()
     val reactionListeners = mutableListOf<(e: MessageReactionAddEvent) -> Unit>()
 
@@ -41,19 +41,19 @@ class Bot(token: String) : ListenerAdapter() {
             Data.log("Bot", "Private message from ${msg.author.asTag} (Channel id: ${msg.channel.id}): $content")
         }
         if (self.jda.presence.status == OnlineStatus.ONLINE) {
-            for (aCommand in adminCommands) {
-                if (content.startsWith(prefix + aCommand.key) && Data.admins.any { it.id == msg.author.id }) {
+            for (aCommand in commands.filter { c -> c.isAdminOnly }) {
+                if (content.startsWith(prefix + aCommand.head) && Data.admins.any { it.id == msg.author.id }) {
                     Data.log("Bot", "Admin command received: $content Author: ${msg.author.asTag} (${msg.author.id})")
-                    msg.delete().queue()
-                    aCommand.value(msg)
+                    if (msg.isFromGuild) msg.delete().queue()
+                    aCommand.action(msg)
                     break
                 }
             }
-            for (command in commands) {
+            for (command in commandsOld) {
                 if (content.startsWith(prefix + command.key)) {
                     msg.channel.sendTyping().queue()
                     Data.log("Bot", "Command received: $content Author: ${msg.author.asTag} (${msg.author.id})")
-                    msg.delete().queue()
+                    if (msg.isFromGuild) msg.delete().queue()
                     command.value(msg)
                     break
                 }
@@ -62,7 +62,7 @@ class Bot(token: String) : ListenerAdapter() {
                 if (content.startsWith(prefix + prefix + cc.command)) {
                     msg.channel.sendTyping().queue()
                     Data.log("Bot", "Custom command received: $content Author: ${msg.author.asTag} (${msg.author.id})")
-                    msg.delete().queue()
+                    if (msg.isFromGuild) msg.delete().queue()
                     msg.channel.sendMessage(cc.output).queue {
                         if (cc.isRemovable) it.makeRemovable()
                     }
